comments Datastore and Implementation, refactors components to reference Datastore
