--- src/scripts/ImplementationStubs.ts
@@ -1,11 +1,0 @@
-/**
- * the following functions must be completed for a functional webapp. Do not change the function names. The function arguments and return values must remain unchanged, otherwise edits to the source code must occur. Any auxilary functions may be written in this file or in an external file.
- */
-
-/**
- * handles user auth state cahnges. the callback must be called whenever the authenticated user changes. The userAuth parameter to the callback function must be an object that contains at least a 'uid' key which contains a string identifying the user
- * @param callback
- */
-export function handleAuthStateChange(callback: (userAuth: any) => void): void {
-  // TODO: complete this stub
-}
