+++ src/store/FireActions.ts
@@ -3,7 +3,7 @@
  */
 
 import { createAction, createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { getUserPrivilege } from "scripts/FireConfig";
+import { getUserPrivilege } from "scripts/Datastore";
 
 export const login_action = createAction("login");
 export const logout_action = createAction("logout");
